requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.2